+++ src/calc-size-fast.h
@@ -1,2 +1,0 @@
-
-int calc_size_fast_outer(SEXP s);