+++ src/init.c
@@ -5,7 +5,6 @@
 extern SEXP marshall_();
 extern SEXP unmarshall_();
 extern SEXP calc_size_robust_();
-extern SEXP calc_size_fast_();
 extern SEXP marshall_fast_();
 
 static const R_CallMethodDef CEntries[] = {
@@ -13,8 +12,7 @@ static const R_CallMethodDef CEntries[] = {
   {"marshall_"        , (DL_FUNC) &marshall_            , 1},
   {"unmarshall_"      , (DL_FUNC) &unmarshall_          , 1},
   {"calc_size_robust_", (DL_FUNC) &calc_size_robust_    , 1},
-  {"calc_size_fast_"  , (DL_FUNC) &calc_size_fast_      , 1},
-  {"marshall_fast_"   , (DL_FUNC) &marshall_fast_       , 2},
+  {"marshall_fast_"   , (DL_FUNC) &marshall_fast_       , 1},
   {NULL , NULL, 0}
 };